Move to the Graphic module and add Read {Graphic, Raster, Color} nodes
