+++ node-graph/nodes/vector/src/instance.rs
@@ -1,5 +1,5 @@
+use core_types::Color;
 use core_types::table::{Table, TableRowRef};
-use core_types::{AnyHash, Color};
 use core_types::{CloneVarArgs, Context, Ctx, ExtractAll, OwnedContextImpl};
 use graphic_types::Graphic;
 use graphic_types::Vector;
@@ -80,39 +80,6 @@ async fn instance_repeat<T: Into<Graphic> + Default + Send + Clone + 'static>(
 	}
 
 	result_table
-}
-
-#[node_macro::node(category("Vector"), path(graphene_core::vector))]
-async fn map<Item: AnyHash + Send + Sync + std::hash::Hash>(
-	ctx: impl Ctx + CloneVarArgs + ExtractAll,
-	#[implementations(
-		Table<Graphic>,
-		Table<Vector>,
-		Table<Raster<CPU>>,
-		Table<Color>,
-		Table<GradientStops>,
-	)]
-	content: Table<Item>,
-	#[implementations(
-		Context -> Table<Graphic>,
-		Context -> Table<Vector>,
-		Context -> Table<Raster<CPU>>,
-		Context -> Table<Color>,
-		Context -> Table<GradientStops>,
-	)]
-	mapped: impl Node<Context<'static>, Output = Table<Item>>,
-) -> Table<Item> {
-	let mut rows = Table::new();
-
-	for (i, row) in content.into_iter().enumerate() {
-		let owned_ctx = OwnedContextImpl::from(ctx.clone());
-		let owned_ctx = owned_ctx.with_vararg(Box::new(Table::new_from_row(row))).with_index(i);
-		let table = mapped.eval(owned_ctx.into_context()).await;
-
-		rows.extend(table);
-	}
-
-	rows
 }
 
 #[cfg(test)]
